Work in progress.

+++ Source/Pages/PledgeConfirm.php
@@ -23,16 +23,43 @@
 				echo("<label>Project: " . $project->name . "</label><br /><br />");
 
 				$userOrganizer = $persistenceClient->userGetById($project->userIDOrganizer);
-				echo("<label>Organizer: " . $userOrganizer->nameFull . "</label><br /><br />");
 
 				$pledgeAmountInUsd = $_POST["AmountToPledgeInUsd"];
-				echo("<label>Amount to Pledge: $" . $pledgeAmountInUsd . "</label><br /><br />");
+
+				if (isset($_POST["IsConfirmed"]) == false || $_POST["IsConfirmed"] == false)
+				{
+					echo("<label>Organizer: " . $userOrganizer->nameFull . "</label><br /><br />");
+				}
+				else
+				{
+					$timePledged = date("Y-m-d H:i:s");
+
+					$pledge = new UserProjectPledge(
+						null, // ID
+						$userLoggedIn->userID,
+						$projectID,
+						$pledgeAmountInUsd,
+						$timePledged,
+						true // $isActive
+					);
+
+					$persistenceClient->userProjectPledgeSave($pledge);
+
+					header("Location: ProjectDetails.php?projectID=" . $projectID);
+				}
+
 		?>
-		
-				<form action="PledgeSave.php" method="post">
+
+				<form action="" method="post">
+					<label>Amount to Pledge: $</label>
+					<input name="AmountToPledgeInUsd" type="number" readonly="true" value="<?php echo($pledgeAmountInUsd) ?>"></input>
+					<br /><br />
+					<label>Activate this checkbox and then click the confirm button to confirm this pledge:</label>
+					<input name="IsConfirmed" type="checkbox"></input>
 					<button type="submit">Confirm</button>
+					<a href="ProjectDetails.php?projectID=<?php echo($projectID) ?>">Cancel</a>
 				</form>
-		
+
 		<?php
 			}
 			else
